Project structure small refactoring
