+++ src/main/java/org/cqfn/astranaut/core/utils/visualizer/DotRender.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cqfn.astranaut.core.utils;
+package org.cqfn.astranaut.core.utils.visualizer;
 
 import java.util.Objects;
 import org.cqfn.astranaut.core.EmptyTree;
